refactor: update transaction type imports to use the new models structure
