+++ apps/mobile/src/hooks/transactions/useTransactionForm.ts
@@ -1,5 +1,6 @@
 import { useCallback, useState } from 'react'
-import { createTransaction, type TransactionType } from '@/services/database/queries/transaction'
+import { createTransaction } from '@/services/database/repository/transaction'
+import { TransactionType } from '@/services/database/models/transaction'
 import { ACCOUNTS, MAX_AMOUNT_CENTS, TRANSACTION_FORM_ERRORS } from '@/utils/transactionForm'
 
 export type TransactionInitialValues = {
